Create midnight-catppuccin-macchiato.theme
Different catppuccin flavour, because I liked the blue-ish tones more. Changed the font to Varela, because I think it looks neat and friendly and matches well with theme. You can adjust the HSLA values, because I kinda did them haphazardly, with the mindset of "it barely changes anything". So yeah,